moved NODES_RELS_PATTERN_REGEX
