+++ wiggle_query_language/clauses/make/parse_make/parse_make_statement.py
@@ -9,6 +9,8 @@ from wiggle_query_language.clauses.parsing_helpers.parse_statement_checks import
 )
 from wiggle_query_language.clauses.regexes.make.make_patterns import (
     MAKE_STATEMENT_ALL_REGEX,
+)
+from wiggle_query_language.clauses.regexes.patterns.patterns import (
     NODES_RELS_PATTERN_REGEX,
 )
